feat: support sparse vector with 1_048_575 dims (#399)
* feat: support sparse vector with 1000000 dims


* change max dims to 1_048_576


* change max dims to 1_048_575


* update PQ dims to u32


* cargo clippy


* update parse function to previous impl


---------

+++ crates/base/src/global/bvecf32_l2.rs
@@ -64,8 +64,8 @@ impl GlobalProductQuantization for BVecf32L2 {
     type ProductQuantizationL2 = BVecf32L2;
 
     fn product_quantization_distance(
-        _dims: u16,
-        _ratio: u16,
+        _dims: u32,
+        _ratio: u32,
         _centroids: &[Scalar<Self>],
         _lhs: Borrowed<'_, Self>,
         _rhs: &[u8],
@@ -74,8 +74,8 @@ impl GlobalProductQuantization for BVecf32L2 {
     }
 
     fn product_quantization_distance2(
-        _dims: u16,
-        _ratio: u16,
+        _dims: u32,
+        _ratio: u32,
         _centroids: &[Scalar<Self>],
         _lhs: &[u8],
         _rhs: &[u8],
@@ -84,8 +84,8 @@ impl GlobalProductQuantization for BVecf32L2 {
     }
 
     fn product_quantization_distance_with_delta(
-        _dims: u16,
-        _ratio: u16,
+        _dims: u32,
+        _ratio: u32,
         _centroids: &[Scalar<Self>],
         _lhs: Borrowed<'_, Self>,
         _rhs: &[u8],
