+++ crates/base/src/global/vecf32_cos.rs
@@ -103,8 +103,8 @@ impl GlobalProductQuantization for Vecf32Cos {
         "aarch64+neon"
     ))]
     fn product_quantization_distance<'a>(
-        dims: u16,
-        ratio: u16,
+        dims: u32,
+        ratio: u32,
         centroids: &[F32],
         lhs: Vecf32Borrowed<'a>,
         rhs: &[u8],
@@ -134,8 +134,8 @@ impl GlobalProductQuantization for Vecf32Cos {
         "aarch64+neon"
     ))]
     fn product_quantization_distance2(
-        dims: u16,
-        ratio: u16,
+        dims: u32,
+        ratio: u32,
         centroids: &[F32],
         lhs: &[u8],
         rhs: &[u8],
@@ -165,8 +165,8 @@ impl GlobalProductQuantization for Vecf32Cos {
         "aarch64+neon"
     ))]
     fn product_quantization_distance_with_delta<'a>(
-        dims: u16,
-        ratio: u16,
+        dims: u32,
+        ratio: u32,
         centroids: &[F32],
         lhs: Vecf32Borrowed<'a>,
         rhs: &[u8],
